requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-crispy-forms==1.14.0
django-crum==0.7.9
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-preventconcurrentlogins==0.8.2
django-resized==1.0.2
django-simple-history==3.1.1
et-xmlfile==1.1.0
idna==3.4
lml==0.1.0
openpyxl==3.0.10
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
